refactor(website): one shared page fragment per file
Three files were doing the same job from two different places: site-footer.ts
and docs-shell.ts sat at lib/*.ts while pageHeader and ctaPanel shared
lib/utils/ui.ts. All four are now one fragment per file under lib/utils/ui/,
mirroring the framework's own one-function-per-file rule for actions and
queries.

lib/ui/ is deliberately NOT the home for these, despite being the shorter
name. It is gitignored (/lib/ui/ in .gitignore) and generated:
scripts/copy-registry.mjs writes the @webjsdev/ui kit's lib there before every
dev and start and prunes anything not upstream. Tracked files put there would
never be committed and would be deleted on the next dev cycle.

Two rounds of moves, two classes of reference the path-alias rewrite could not
see: a test that reads a file by literal path, and a dozen doc comments naming
the old location. Both only surfaced under a full test run, which is the
argument for running one after a move rather than trusting a grep.

35 tests pass, webjs check clean, all eight routes 200.
