+++ website/app/layout.ts
@@ -1,7 +1,7 @@
 import { html, cspNonce } from '@webjsdev/core';
 import '#components/theme-toggle.ts';
 import { DOCS_START_PATH, UI_PATH, EXAMPLE_BLOG_URL, GH_URL, NEW_TAB } from '#lib/links.ts';
-import { siteFooter } from '#lib/site-footer.ts';
+import { siteFooter } from '#lib/utils/ui/site-footer.ts';
 import { brandLockup } from '#lib/brand.ts';
 
 /**
@@ -149,7 +149,7 @@ export default function RootLayout({ children }: { children: unknown }) {
         if (document.readyState === 'loading') document.addEventListener('DOMContentLoaded', measure);
         else measure();
       })();
-      // The sidebar drawer shared by /docs and /ui (lib/docs-shell.ts) is
+      // The sidebar drawer shared by /docs and /ui (lib/utils/ui/docs-shell.ts) is
       // driven entirely from here, and both halves of that are deliberate.
       //
       // Its open state is an attribute on <body>, which is OUTSIDE every swap
